Minor shuffle around and cleanup.
This just makes #363 a  bit easier to read.
